requirements.txt
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.2
django-heroku==0.3.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
drf-yasg==1.20.0
gunicorn==20.1.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
packaging==20.9
Pillow==8.3.1
psycopg2==2.9.1
pyforest==1.0.3
PyJWT==2.3.0 
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
ruamel.yaml==0.17.7
ruamel.yaml.clib==0.2.2
six==1.16.0
sqlparse==0.4.1

typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.5
whitenoise==5.3.0
wincertstore==0.2
